aio-checker-pro/real-estate-app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
pillow==10.1.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.2
PyYAML==6.0.1